+++ api/src/main/kotlin/glimpse/Matrices.kt
@@ -44,7 +44,7 @@ fun orthographicProjectionMatrix(left: Float, right: Float, bottom: Float, top: 
 	return Matrix(listOf(
 			2f / width, 0f, 0f, 0f,
 			0f, 2f / height, 0f, 0f,
-			0f, 0f, 2f / depth, 0f,
+			0f, 0f, -2f / depth, 0f,
 			-(right + left) / width, -(top + bottom) / height, -(near + far) / depth, 1f))
 }
 
@@ -138,11 +138,11 @@ fun scalingMatrix(scale: Float): Matrix = scalingMatrix(scale, scale, scale)
 /**
  * Returns a transformation [Matrix] for scaling.
  */
-fun scalingMatrix(scaleX: Float, scaleY: Float, scaleZ: Float): Matrix =
+fun scalingMatrix(x: Float = 1f, y: Float = 1f, z: Float = 1f): Matrix =
 		Matrix(listOf(
-				scaleX, 0f, 0f, 0f,
-				0f, scaleY, 0f, 0f,
-				0f, 0f, scaleZ, 0f,
+				x, 0f, 0f, 0f,
+				0f, y, 0f, 0f,
+				0f, 0f, z, 0f,
 				0f, 0f, 0f, 1f))
 
 /**