+++ jogl/src/main/kotlin/glimpse/jogl/gles/GLES.kt
@@ -1,5 +1,6 @@
 package glimpse.jogl.gles
 
+import com.jogamp.opengl.GL2ES2
 import com.jogamp.opengl.util.texture.TextureIO
 import glimpse.Color
 import glimpse.gles.*
@@ -17,7 +18,6 @@ import java.nio.Buffer
 import java.nio.ByteBuffer
 import java.nio.FloatBuffer
 import java.nio.IntBuffer
-import javax.media.opengl.GL2ES2
 
 /**
  * JOGL implementation of GLES facade.
@@ -137,16 +137,24 @@ class GLES(val gles: GL2ES2) : GLES {
 		gles.glUniform1f(location.value, float)
 	}
 
+	override fun uniformFloats(location: UniformLocation, floats: FloatArray) {
+		gles.glUniform1fv(location.value, floats.size, floats, 0)
+	}
+
 	override fun uniformInt(location: UniformLocation, int: Int) {
 		gles.glUniform1i(location.value, int)
+	}
+
+	override fun uniformInts(location: UniformLocation, ints: IntArray) {
+		gles.glUniform1iv(location.value, ints.size, ints, 0)
 	}
 
 	override fun uniformMatrix16f(location: UniformLocation, _16f: Array<Float>) {
 		gles.glUniformMatrix4fv(location.value, 1, false, _16f.toFloatArray(), 0)
 	}
 
-	override fun uniform4f(location: UniformLocation, _4f: Array<Float>) {
-		gles.glUniform4fv(location.value, 1, _4f.toFloatArray(), 0)
+	override fun uniform4f(location: UniformLocation, _4f: Array<Float>, count: Int) {
+		gles.glUniform4fv(location.value, count, _4f.toFloatArray(), 0)
 	}
 
 	override fun createAttributeFloatArray(location: AttributeLocation, buffer: FloatBuffer, vectorSize: Int) =